+++ libs/innovatech/ui/guarantees/src/lib/services/guarantee-data.service.ts
@@ -1,4 +1,4 @@
-import { HttpClient } from '@angular/common/http';
+import { HttpClient, HttpParams } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { ApsDataService, saveFile } from '@arphase/ui/core';
 import { Guarantee } from '@innovatech/common/domain';
@@ -15,10 +15,9 @@ export class GuaranteeDataService extends ApsDataService<Guarantee> {
   }
 
   getGuaranteePdf(id: number): Observable<Blob> {
+    const params = new HttpParams({ fromObject: { utcOffset: -new Date().getTimezoneOffset() } });
     return this.http
-      .get(`/ivtApi/guarantees/export/pdf/${id}`, {
-        responseType: 'blob',
-      })
+      .get(`/ivtApi/guarantees/export/pdf/${id}`, { responseType: 'blob', params })
       .pipe(tap((file: Blob) => saveFile(file, `Garantía ${id}.pdf`)));
   }
 }